feat(threads): 创建/编辑主题时传入了价格，必须传附件。

+++ app/Commands/Thread/CreateThread.php
@@ -153,7 +153,7 @@ class CreateThread
         $thread->location = Arr::get($this->data, 'attributes.location', '');
 
         // 附件价格
-        $thread->attachment_price = Arr::get($this->data, 'attributes.attachment_price', 0);
+        $thread->attachment_price = (float) Arr::get($this->data, 'attributes.attachment_price', 0);
 
         $thread->setRelation('user', $this->actor);
